+++ LiteDB.Benchmarks/Benchmarks/Generator/DataGenerationBenchmark.cs
@@ -6,7 +6,7 @@ using LiteDB.Benchmarks.Models.Generators;
 
 namespace LiteDB.Benchmarks.Benchmarks.Generator
 {
-    [BenchmarkCategory(LiteDB.Benchmarks.Constants.Categories.DATA_GEN)]
+    [BenchmarkCategory(Constants.Categories.DATA_GEN)]
     public class DataGenerationDatabaseBenchmark : BenchmarkBase
     {
         [Benchmark]
