+++ LiteDB.Benchmarks/Benchmarks/Insertion/InsertionBasicBenchmark.cs
@@ -8,10 +8,10 @@ using LiteDB.Benchmarks.Models.Generators;
 
 namespace LiteDB.Benchmarks.Benchmarks.Insertion
 {
-    [BenchmarkCategory(LiteDB.Benchmarks.Constants.Categories.INSERTION)]
+    [BenchmarkCategory(Constants.Categories.INSERTION)]
     public class InsertionBasicBenchmark : DatabaseBenchmarkBase
     {
-        protected override string DatabasePath => Constants.DatabaseNames.Insertion;
+        protected override string DatabasePath => Constants.DatabaseNames.INSERTION;
 
         private List<FileMetaBase> data;
