--- LiteDB.Benchmarks/Constants.cs
@@ -1,13 +1,0 @@
-namespace LiteDB.Benchmarks
-{
-    internal static class Constants
-    {
-        internal static class Categories
-        {
-            internal const string DATA_GEN = nameof(DATA_GEN);
-            internal const string QUERIES = nameof(QUERIES);
-            internal const string INSERTION = nameof(INSERTION);
-            internal const string DELETION = nameof(DELETION);
-        }
-    }
-}
